Begin Peripheral 3: KeyPad Driver. Not working in sim.
Running the sim, check dOut @ 25ns. It should take on the value of 4, but it doesn't -- it has a value of 1. This makes no sense, given that the input to the DReg is *never* 1 in this sim. Also, it updates on the 'LD' posedge, which doesn't make sense given that the dOut should update one clock cycle *after* ld goes high.